# architecture/api-contract-openapi.yaml
openapi: 3.0.3
info:
  title: iReg Moto BF API
  description: >
    API réglementaire pour la conformité des deux-roues motorisés au Burkina Faso.
    Plateforme SaaS interfaçant importateurs, distributeurs, unités d'assemblage,
    détaillants et le ministère des transports (DRCTT).
    Supporte REST (mutations métier) et GraphQL (analytics).
  version: 1.0.0
  contact:
    name: iReg Moto BF Technical Team
    email: tech@ireg-moto.bf
  license:
    name: Proprietary
    url: https://ireg-moto.bf/licence
servers:
  - url: https://api.ireg-moto.bf/v1
    description: Production
  - url: https://api-staging.ireg-moto.bf/v1
    description: Staging
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authentification par email/mot de passe
      description: >
        Authentifie un utilisateur et retourne un JWT access token.
        Si MFA est requis pour ce profil, retourne un token temporaire
        et nécessite une vérification MFA.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentification réussie (ou MFA requis)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Compte verrouillé ou suspendu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchissement du token d'accès
      description: >
        Échange un refresh token valide contre un nouveau couple
        access token + refresh token (rotation sécurisée).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: Refresh token JWT
                  example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Nouveaux tokens générés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/mfa/setup:
    post:
      tags: [Auth]
      summary: Configuration MFA (TOTP)
      description: >
        Génère un secret TOTP et un QR code pour l'enregistrement
        dans une application d'authentification (Google Authenticator, Authy).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfaType]
              properties:
                mfaType:
                  type: string
                  enum: [TOTP, SMS]
                  example: TOTP
      responses:
        '200':
          description: Secret MFA généré
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                    example: JBSWY3DPEHPK3PXP
                  qrCodeUrl:
                    type: string
                    format: uri
                    example: otpauth://totp/iRegMoto:user@example.com?secret=JBSWY3DPEHPK3PXP&issuer=iRegMoto
                  backupCodes:
                    type: array
                    items:
                      type: string
                    example: ["12345678", "87654321", "11223344"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Vérification MFA
      description: >
        Vérifie un code TOTP ou SMS et finalise l'authentification.
        Retourne les tokens JWT définitifs après validation MFA.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tempToken, mfaCode]
              properties:
                tempToken:
                  type: string
                  description: Token temporaire retourné par /auth/login
                mfaCode:
                  type: string
                  pattern: '^[0-9]{6}$'
                  description: Code TOTP à 6 chiffres
                  example: "123456"
                mfaType:
                  type: string
                  enum: [TOTP, SMS, BACKUP_CODE]
                  default: TOTP
      responses:
        '200':
          description: MFA validé, authentification complète
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Code MFA invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/oauth/callback:
    post:
      tags: [Auth]
      summary: Callback OAuth2 (Google, Microsoft)
      description: >
        Échange un code d'autorisation OAuth2 contre un compte iReg.
        Crée le compte s'il n'existe pas (first login).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, code]
              properties:
                provider:
                  type: string
                  enum: [GOOGLE, MICROSOFT]
                code:
                  type: string
                  description: Code d'autorisation OAuth2
                redirectUri:
                  type: string
                  format: uri
      responses:
        '200':
          description: Authentification OAuth réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Déconnexion
      description: >
        Révoque le token d'accès courant et ajoute le refresh token
        à la liste noire Redis.
      responses:
        '204':
          description: Déconnexion réussie
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Profil utilisateur connecté
      description: Retourne les informations du utilisateur authentifié.
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================================================
  # ACTEURS (MODULE A)
  # ==========================================================================
  /actors:
    get:
      tags: [Acteurs]
      summary: Liste des acteurs économiques
      description: >
        Retourne la liste paginée des acteurs économiques avec filtres.
        Un importateur ne voit que son propre acteur. Un admin voit tous.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, SUSPENDED, REVOKED, EXPIRED]
        - name: actorType
          in: query
          schema:
            type: string
            enum: [IMPORTATEUR, DISTRIBUTEUR, ASSEMBLEUR, DETAILLANT, PRESTATAIRE]
        - name: region
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
          description: Recherche par nom, NIF, RCCM
        - name: complianceScoreMin
          in: query
          schema: { type: number, minimum: 0, maximum: 100 }
        - name: agreementStatus
          in: query
          schema:
            type: string
            enum: [DRAFT, SUBMITTED, UNDER_REVIEW, APPROVED, REJECTED, EXPIRED]
      responses:
        '200':
          description: Liste des acteurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActors'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Acteurs]
      summary: Créer un acteur économique
      description: >
        Crée un nouvel acteur économique avec statut PENDING.
        Déclenche le workflow d'agrément.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActorRequest'
      responses:
        '201':
          description: Acteur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: NIF ou email déjà existant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /actors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Acteurs]
      summary: Détail d'un acteur
      responses:
        '200':
          description: Détail acteur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Acteurs]
      summary: Modifier un acteur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActorRequest'
      responses:
        '200':
          description: Acteur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /actors/{id}/documents:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Acteurs]
      summary: Documents d'un acteur
      responses:
        '200':
          description: Liste des documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActorDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Acteurs]
      summary: Ajouter un document à un acteur
      description: Upload d'un document KYC (multipart/form-data).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, documentType]
              properties:
                file:
                  type: string
                  format: binary
                documentType:
                  type: string
                  enum: [REGISTRE_COMMERCE, NIF, ASSURANCE, CERTIFICATE_ORIGIN, CONTRAT_DISTRIBUTION, CERT_AGREEMENT, LOCAL_CONTENT_PLAN]
                documentLabel:
                  type: string
                issuedDate:
                  type: string
                  format: date
                expiryDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Document ajouté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDocument'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Fichier trop volumineux (max 10MB)
  /actors/{id}/warehouses:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Acteurs]
      summary: Entrepôts d'un acteur
      responses:
        '200':
          description: Liste des entrepôts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Warehouse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Acteurs]
      summary: Créer un entrepôt pour un acteur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWarehouseRequest'
      responses:
        '201':
          description: Entrepôt créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT d'accès (RS256)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API pour les intégrations système (ministère, DGI)

  # ========================================================================
  # RESPONSES
  # ========================================================================
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Non authentifié — Token invalide ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Accès interdit — Permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Trop de requêtes — Rate limiting
      headers:
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ========================================================================
  # SCHEMAS
  # ========================================================================
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Le VIN fourni n'est pas valide"
        code:
          type: string
          example: INVALID_VIN
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /api/v1/vehicles
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: contact@societe-import.bf
        password:
          type: string
          format: password
          minLength: 8
          example: ""
        deviceId:
          type: string
          description: Identifiant unique du device (PWA)
        deviceInfo:
          type: string
    LoginResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        mfaRequired:
          type: boolean
          description: true si MFA requis pour compléter l'auth
        tempToken:
          type: string
          description: Token temporaire pour vérification MFA
        accessToken:
          type: string
          description: JWT access token (si pas de MFA requis)
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Durée de validité access token (secondes)
          example: 900
        user:
          $ref: '#/components/schemas/User'
    TokenRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [SUPER_ADMIN, ADMIN_DRCTT, CONTROLEUR, IMPORTATEUR, DISTRIBUTEUR, ASSEMBLEUR, DETAILLANT, AUDITEUR, CNTI_AGENT]
              actorId:
                type: string
        permissions:
          type: array
          items:
            type: string
        locale:
          type: string
          example: fr_BF
        mfaEnabled:
          type: boolean
        actor:
          type: object
          properties:
            id:
              type: string
            companyName:
              type: string
            actorType:
              type: string

    # -------------------------------------------------------------------------
    # Acteurs
    # -------------------------------------------------------------------------
    Actor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actorType:
          type: string
          enum: [IMPORTATEUR, DISTRIBUTEUR, ASSEMBLEUR, DETAILLANT, PRESTATAIRE]
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED, REVOKED, EXPIRED]
        companyName:
          type: string
        tradeName:
          type: string
        nif:
          type: string
        rccm:
          type: string
        legalRepresentativeName:
          type: string
        email:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        region:
          type: string
        agreementStatus:
          type: string
        agreementNumber:
          type: string
        agreementExpiresAt:
          type: string
          format: date-time
        complianceScore:
          type: number
        complianceCountdownDays:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ActorDetail:
      allOf:
        - $ref: '#/components/schemas/Actor'
        - type: object
          properties:
            warehouses:
              type: array
              items:
                $ref: '#/components/schemas/Warehouse'
            documents:
              type: array
              items:
                $ref: '#/components/schemas/ActorDocument'
            parentActor:
              type: object
              properties:
                id:
                  type: string
                companyName:
                  type: string
    PaginatedActors:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Actor'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateActorRequest:
      type: object
      required: [actorType, companyName, legalRepresentativeName, legalRepresentativePhone, email, phone, addressLine1, city, region, nif]
      properties:
        actorType:
          type: string
          enum: [IMPORTATEUR, DISTRIBUTEUR, ASSEMBLEUR, DETAILLANT, PRESTATAIRE]
        companyName:
          type: string
          maxLength: 255
        tradeName:
          type: string
        nif:
          type: string
        rccm:
          type: string
        legalRepresentativeName:
          type: string
        legalRepresentativeTitle:
          type: string
        legalRepresentativePhone:
          type: string
        legalRepresentativeEmail:
          type: string
        email:
          type: string
        phone:
          type: string
        alternatePhone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        region:
          type: string
        postalCode:
          type: string
        gpsCoordinates:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        bankName:
          type: string
        bankAccount:
          type: string
        bankRib:
          type: string
        parentActorId:
          type: string
          format: uuid
    UpdateActorRequest:
      type: object
      properties:
        companyName: { type: string }
        tradeName: { type: string }
        legalRepresentativeName: { type: string }
        legalRepresentativePhone: { type: string }
        email: { type: string }
        phone: { type: string }
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        region: { type: string }
        gpsCoordinates:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        bankName: { type: string }
        bankAccount: { type: string }
    ActorDocument:
      type: object
      properties:
        id:
          type: string
        documentType:
          type: string
        documentLabel:
          type: string
        fileName:
          type: string
        fileSizeBytes:
          type: integer
        mimeType:
          type: string
        storageUrl:
          type: string
        checksumSha256:
          type: string
        issuedDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        isVerified:
          type: boolean
        verifiedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Warehouse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        region:
          type: string
        gpsCoordinates:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        isPrimary:
          type: boolean
        capacityUnits:
          type: integer
        surfaceSqm:
          type: number
        securityLevel:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateWarehouseRequest:
      type: object
      required: [name, code, addressLine1, city, region]
      properties:
        name: { type: string }
        code: { type: string }
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        region: { type: string }
        gpsCoordinates:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        capacityUnits: { type: integer }
        surfaceSqm: { type: number }
        securityLevel:
          type: string
          enum: [STANDARD, RENFORCE, HAUTE]
        isPrimary: { type: boolean }

    # -------------------------------------------------------------------------
    # Véhicules
    # -------------------------------------------------------------------------
...